Revert moving caching logic and fix type gaurd errors
